+++ inst/NEWS.md
@@ -1,10 +1,9 @@
 # CHANGES IN VERSION 2.4.15 (In development)
 ## BUG REPORTS
 
-##ENHANCEMENTS
+## ENHANCEMENTS
+- Added `sampleOrder1` and `sampleOrder2` arguments to `coOncoplot`. Issue: [592](https://github.com/PoisonAlien/maftools/issues/592)
 - Added `filterMaf` function. 
-
-
 
 # CHANGES IN VERSION 2.4.10 (Release/Bioconductor version)
 ## BUG FIXES